requirements.txt
python-telegram-bot[webhooks]==20.5
httpx==0.27.0